move cli mod to config mod
