+++ src/config/mod.rs
@@ -1,3 +1,4 @@
+pub mod cli;
 mod client;
 mod server;
